# monserverspirng/src/main/java/org/openapitools/api/BooksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Book;
import org.openapitools.model.Error;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-10-06T15:19:12.620607+02:00[Europe/Berlin]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "books", description = "API de gestion des livres")
public interface BooksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_BOOKS_ID_DELETE = "/books/{id}";
    /**
     * DELETE /books/{id} : Suppression d&#39;un livres
     *
     * @param id  (required)
     * @return Le livre a été supprimé (status code 204)
     *         or livre non trouvé (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "booksIdDelete",
        summary = "Suppression d'un livres",
        tags = { "books" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Le livre a été supprimé"),
            @ApiResponse(responseCode = "404", description = "livre non trouvé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = BooksApi.PATH_BOOKS_ID_DELETE,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> booksIdDelete(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKS_ID_GET = "/books/{id}";
    /**
     * GET /books/{id} : Retourner un livre au moyen de son ID
     *
     * @param id  (required)
     * @return Le livre est retourné avec succès (status code 200)
     *         or livre non trouvé (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "booksIdGet",
        summary = "Retourner un livre au moyen de son ID",
        tags = { "books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Le livre est retourné avec succès", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "404", description = "livre non trouvé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_BOOKS_ID_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<Book> booksIdGet(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"publish\" : true, \"description\" : \"Lorem ipsum\", \"id\" : 10, \"title\" : \"La maison du guet\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKS_ID_PUT = "/books/{id}";
    /**
     * PUT /books/{id} : Mise à jour des information d&#39;un livres
     *
     * @param id  (required)
     * @param book  (required)
     * @return Le livre est maj avec succès (status code 200)
     *         or livre non trouvé (status code 404)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "booksIdPut",
        summary = "Mise à jour des information d'un livres",
        tags = { "books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Le livre est maj avec succès", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            }),
            @ApiResponse(responseCode = "404", description = "livre non trouvé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = BooksApi.PATH_BOOKS_ID_PUT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Book> booksIdPut(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"publish\" : true, \"description\" : \"Lorem ipsum\", \"id\" : 10, \"title\" : \"La maison du guet\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKS_POST = "/books";
    /**
     * POST /books : Ajouter un livre.
     *
     * @param book  (required)
     * @return Livre crée avec succès (status code 201)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "booksPost",
        summary = "Ajouter un livre.",
        tags = { "books" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Livre crée avec succès"),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BooksApi.PATH_BOOKS_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> booksPost(
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_BOOKS = "/books";
    /**
     * GET /books : Retourner des livres
     * Retourner tous les livres en base de données.
     *
     * @return Les livres sont retournés avec succès. (status code 200)
     *         or Internal Server Error (status code 500)
     */
    @Operation(
        operationId = "getBooks",
        summary = "Retourner des livres",
        description = "Retourner tous les livres en base de données.",
        tags = { "books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Les livres sont retournés avec succès.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Book.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Internal Server Error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_GET_BOOKS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Book>> getBooks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"publish\" : true, \"description\" : \"Lorem ipsum\", \"id\" : 10, \"title\" : \"La maison du guet\" }, { \"publish\" : true, \"description\" : \"Lorem ipsum\", \"id\" : 10, \"title\" : \"La maison du guet\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"[\\"Aucun livre trouvé.\\",\\"Erreur interne de serveur\\"]\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}