# monserverspirng/src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  description: This is a simple REST API perfoming CRUD operations
  title: Book API
  version: 1.0.0
servers:
- description: SERVER DEV
  url: http:localhost:8080/api/v1
- description: SERVER PPROD
  url: http:localhost:8081/api/v1
tags:
- description: API de gestion des livres
  name: books
paths:
  /books:
    get:
      description: Retourner tous les livres en base de données.
      operationId: getBooks
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Book"
                type: array
          description: Les livres sont retournés avec succès.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      summary: Retourner des livres
      tags:
      - books
      x-accepts:
      - application/json
      x-tags:
      - tag: books
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "201":
          description: Livre crée avec succès
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      summary: Ajouter un livre.
      tags:
      - books
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: books
  /books/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Le livre a été supprimé
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: livre non trouvé
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      summary: Suppression d'un livres
      tags:
      - books
      x-accepts:
      - application/json
      x-tags:
      - tag: books
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
          description: Le livre est retourné avec succès
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: livre non trouvé
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      summary: Retourner un livre au moyen de son ID
      tags:
      - books
      x-accepts:
      - application/json
      x-tags:
      - tag: books
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
          description: Le livre est maj avec succès
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: livre non trouvé
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      summary: Mise à jour des information d'un livres
      tags:
      - books
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: books
components:
  schemas:
    Book:
      example:
        publish: true
        description: Lorem ipsum
        id: 10
        title: La maison du guet
      properties:
        id:
          example: 10
          type: integer
        title:
          example: La maison du guet
          type: string
        description:
          example: Lorem ipsum
          type: string
        publish:
          example: true
          type: boolean
      type: object
    Error:
      example:
        code: 0
        message: "[\"Aucun livre trouvé.\",\"Erreur interne de serveur\"]"
      properties:
        code:
          type: integer
        message:
          example: "[\"Aucun livre trouvé.\",\"Erreur interne de serveur\"]"
          type: string
      type: object